# VETERINARIAPROD/gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: VetManager Pro API
  description: |
    API REST para el sistema veterinario SaaS multinacional VetManager Pro.

    **Autenticación:** Bearer JWT (access token de 15min + refresh token de 7 días)

    **Multi-tenant:** Todos los endpoints filtran datos por `branchId` del token JWT.

    **RBAC:** Cada endpoint requiere un permiso específico según el rol del usuario.
  version: "1.0.0"
  contact:
    name: VetManager Pro
    email: api@vetmanagerpro.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.vetmanagerpro.com/api/v1
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  # ── AUTH ────────────────────────────────────────────────────────────────────
paths:
...
  /reports/charts/revenue-trend:
    get:
      tags: [Reports]
      summary: Gráfico de tendencia de ingresos (datasets para Chart.js)
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: groupBy
          schema: { type: string, enum: [day, week, month], default: month }
      responses:
        '200': { description: 'Labels + datasets (gross, net, collected)' }
  /reports/charts/appointments-heatmap:
    get:
      tags: [Reports]
      summary: Heatmap de turnos por día de semana y hora
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200': { description: 'Matriz de calor por dayOfWeek y hora' }
...
  /reports/charts/top-diagnoses-bar:
    get:
      tags: [Reports]
      summary: Top 10 diagnósticos más frecuentes (bar chart)
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200': { description: Diagnósticos por frecuencia }
  /reports/charts/payment-methods:
    get:
      tags: [Reports]
      summary: Distribución de métodos de pago (pie/bar chart)
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200': { description: Pagos por método }
...
  /notifications:
    get:
      tags: [Notifications]
      summary: Listar notificaciones del usuario
      parameters:
        - in: query
          name: unreadOnly
          required: false
          schema: { type: string, enum: ['true', 'false'] }
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: Notificaciones }
...
  /notifications/reminders:
    get:
      tags: [Notifications]
      summary: Listar recordatorios
      parameters:
        - in: query
          name: status
          required: false
          schema: { type: string }
        - in: query
          name: type
          required: false
          schema: { type: string }
        - in: query
          name: page
          required: false
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: Recordatorios }
...
  /notifications/messages/reminder/{reminderId}:
    post:
      tags: [Notifications]
      summary: Enviar mensaje para recordatorio
      parameters:
        - in: path
          name: reminderId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Mensaje de recordatorio enviado }
...
  /portal/notifications:
    get:
      tags: [Portal]
      summary: Notificaciones del dueño
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200': { description: Notificaciones }
...